refactoring and mongodb database strategy
