+++ models/Chatbot.js
@@ -1,5 +1,3 @@
-
-
 // Chatbot prototype
 function Chatbot(strategy) {
     this.chatbotAi = strategy || "";
@@ -11,15 +9,7 @@ Chatbot.prototype = {
     },
     getIntent: function (message) {
         return this.chatbotAi.getIntent(message);
-    },
-    
-    getResponse: function(intent){
-        return this.chatbotAi.getResponse(intent)
-    },
-    setResponse: function(intent,response){
-        return this.chatbotAi.setResponse(intent,response);
     }
-    
 };
 
 module.exports = Chatbot;
